Raise docker client version to easy dev setup on mac os
Moved the docker exception to a different package.
Reimplemented the MockedLogStream.
Move volumes to appropriate spot in new api version

The version bump resulted in a bunch of dependency conflicts.
Ranging from guava to jackson versions.

In the end the solution is the shaded spotify jar,
with an updated version of rest easy that excludes a few
conflicting dependencies.